LPS-110153 Move files
